+++ asyncpraw/models/reddit/mixins/inboxable.py
@@ -103,33 +103,6 @@ class InboxableMixin:
         """
         await self._reddit.inbox.mark_unread([self])
 
-    async def unblock_subreddit(self) -> None:
-        """Unblock a subreddit.
-
-        .. note::
-
-            This method pertains only to objects which were retrieved via the inbox.
-
-        For example, to unblock all blocked subreddits that you can find by going
-        through your inbox:
-
-        .. code-block:: python
-
-            from asyncpraw.models import SubredditMessage
-
-            subs = set()
-            async for item in reddit.inbox.messages(limit=None):
-                if isinstance(item, SubredditMessage):
-                    if (
-                        item.subject == "[message from blocked subreddit]"
-                        and str(item.subreddit) not in subs
-                    ):
-                        item.unblock_subreddit()
-                        subs.add(str(item.subreddit))
-
-        """
-        await self._reddit.post(API_PATH["unblock_subreddit"], data={"id": self.fullname})
-
     async def uncollapse(self) -> None:
         """Mark the item as uncollapsed.